refactor: separate error handling routes
Move error handling routes to a separate route builder.
Thanks to common Camel Context routing between route builder classes
works like a charm.

EVNT-601
