+++ splunk-quarkus/src/main/java/com/redhat/console/notifications/splunkintegration/SplunkIntegration.java
@@ -24,7 +24,6 @@ import javax.enterprise.context.ApplicationScoped;
 
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import org.apache.camel.Exchange;
-import org.apache.camel.LoggingLevel;
 import org.apache.camel.Processor;
 import org.apache.camel.builder.endpoint.EndpointRouteBuilder;
 import org.apache.camel.component.http.HttpClientConfigurer;
@@ -101,10 +100,6 @@ public class SplunkIntegration extends EndpointRouteBuilder {
 
         configureErrorHandler();
         configureIngress();
-        configureIoFailed();
-        configureHttpFailed();
-        configureTargetUrlValidationFailed();
-        configureSecureConnectionFailed();
         configureReturn();
         configureSuccessHandler();
         configureHandler();
@@ -124,76 +119,6 @@ public class SplunkIntegration extends EndpointRouteBuilder {
         onException(ProtocolException.class)
                 .to(direct("secureConnectionFailed"))
                 .handled(true);
-    }
-
-    private void configureSecureConnectionFailed() throws Exception {
-        Processor ceEncoder = new CloudEventEncoder(COMPONENT_NAME, RETURN_TYPE);
-        Processor resultTransformer = new ResultTransformer();
-        // The error handler when we receive an HTTP (unsecure) connection instead of HTTPS
-        from(direct("secureConnectionFailed"))
-                .routeId("secureConnectionFailed")
-                .log(LoggingLevel.ERROR, "ProtocolException for event ${header.ce-id} (orgId ${header.orgId}"
-                                         + " account ${header.accountId}) to ${header.targetUrl}: ${exception.message}")
-                .log(LoggingLevel.DEBUG, "${exception.stacktrace}")
-                .setBody(simple("${exception.message}"))
-                .setHeader("outcome-fail", simple("true"))
-                .process(resultTransformer)
-                .marshal().json()
-                .process(ceEncoder)
-                .to(direct("return"));
-    }
-
-    private void configureTargetUrlValidationFailed() throws Exception {
-        Processor ceEncoder = new CloudEventEncoder(COMPONENT_NAME, RETURN_TYPE);
-        Processor resultTransformer = new ResultTransformer();
-        // The error handler when we receive a TargetUrlValidator failure
-        from(direct("targetUrlValidationFailed"))
-                .routeId("targetUrlValidationFailed")
-                .log(LoggingLevel.ERROR, "IllegalArgumentException for event ${header.ce-id} (orgId ${header.orgId}"
-                                         + " account ${header.accountId}) to ${header.targetUrl}: ${exception.message}")
-                .log(LoggingLevel.DEBUG, "${exception.stacktrace}")
-                .setBody(simple("${exception.message}"))
-                .setHeader("outcome-fail", simple("true"))
-                .process(resultTransformer)
-                .marshal().json()
-                .process(ceEncoder)
-                .to(direct("return"));
-    }
-
-    private void configureIoFailed() throws Exception {
-        Processor ceEncoder = new CloudEventEncoder(COMPONENT_NAME, RETURN_TYPE);
-        Processor resultTransformer = new ResultTransformer();
-        // The error handler found an IO Exception. We set the outcome to fail and then send to kafka
-        from(direct("ioFailed"))
-                .routeId("ioFailed")
-                .log(LoggingLevel.ERROR, "IOFailure for event ${header.ce-id} (orgId ${header.orgId}"
-                                         + " account ${header.accountId}) to ${header.targetUrl}: ${exception.message}")
-                .log(LoggingLevel.DEBUG, "${exception.stacktrace}")
-                .setBody(simple("${exception.message}"))
-                .setHeader("outcome-fail", simple("true"))
-                .process(resultTransformer)
-                .marshal().json()
-                .process(ceEncoder)
-                .to(direct("return"));
-    }
-
-    private void configureHttpFailed() throws Exception {
-        Processor ceEncoder = new CloudEventEncoder(COMPONENT_NAME, RETURN_TYPE);
-        Processor resultTransformer = new ResultTransformer();
-        // The error handler found an HTTP Exception. We set the outcome to fail and then send to kafka
-        from(direct("httpFailed"))
-                .routeId("httpFailed")
-                .log(LoggingLevel.ERROR, "HTTPFailure for event ${header.ce-id} (orgId ${header.orgId} account"
-                                         + " ${header.accountId}) to ${header.targetUrl}: ${exception.getStatusCode()}"
-                                         + " ${exception.getStatusText()}: ${exception.message}")
-                .log(LoggingLevel.DEBUG, "Response Body: ${exception.getResponseBody()}")
-                .log(LoggingLevel.DEBUG, "Response Headers: ${exception.getResponseHeaders()}")
-                .setBody(simple("${exception.message}"))
-                .setHeader("outcome-fail", simple("true"))
-                .process(resultTransformer)
-                .marshal().json()
-                .process(ceEncoder)
-                .to(direct("return"));
     }
 
     private void configureIngress() throws Exception {
